requirements.txt
flask
pyodbc
folium
requests
pillow == 10.2.0
azure-storage-blob == 12.19.1
pymongo == 4.6.3
google-generativeai == 0.5.0